Unreviewed. Remove incorrect assert after r183189.
In this case the assert was already there, but it's incorrect
because we are handling uninitialized attachments below in the
same function, and also when sending messages. So, if it's ok to
send uninitialized attachments, it's ok to receive them. This
didn't fail before r183189, because we were only invaliding the
attachments by setting the file descriptor to -1, and they were
handled as null attachments. After r183189, the move operator
resets the moved attachment leaving it uninitialized again.

* Platform/IPC/unix/ConnectionUnix.cpp:
(IPC::Connection::processMessage):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183258 268f45cc-cd09-0410-ab3c-d52691b4dbfc