theverge.com burning 100% CPU on iPad (due to window.screenX/Y) <https://webkit.org/b/144175> <rdar://problem/20694585>
Reviewed by Anders Carlsson.

Return an empty window rect on iOS WK2 builds. This avoids sending a synchronous
IPC request to the UI process that would give the same result, but spend way
more time and CPU cycles to do so.

With this change, theverge.com CPU usage goes from 100% to ~20% when idle.

* WebProcess/WebCoreSupport/WebChromeClient.cpp:
(WebKit::WebChromeClient::windowRect):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183299 268f45cc-cd09-0410-ab3c-d52691b4dbfc