mayExit() is wrong about Branch nodes with ObjectOrOtherUse: they can exit. https://bugs.webkit.org/show_bug.cgi?id=144152
Reviewed by Filip Pizlo.

Changed the EdgeMayExit functor to recognize ObjectUse, ObjectOrOtherUse,
StringObjectUse, and StringOrStringObjectUse kinds as potentially triggering
OSR exits.  This was overlooked in the original code.

While only the ObjectOrOtherUse kind is relevant for manifesting this bug with
the Branch node, the other 3 may also trigger the same bug for other nodes.
To prevent this bug from manifesting with other nodes (and future ones that
are yet to be added to mayExits()'s "potential won't exit" set), we fix the
EdgeMayExit functor to handle all 4 use kinds (instead of just ObjectOrOtherUse).

Also added a test to exercise a code path that will trigger this bug with
the Branch node before the fix is applied.

* dfg/DFGMayExit.cpp:
* tests/stress/branch-may-exit-due-to-object-or-other-use-kind.js: Added.
(inlinedFunction):
(foo):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183310 268f45cc-cd09-0410-ab3c-d52691b4dbfc