Unreviewed. Fix /webkit2/WebKitWebInspectorServer/test-open-debugging-session.
Latest changes in the inspector UI made this test fail. We were
relying that the sidebar was always visible and the first 'li'
element of the inspector was the one containing the title. This is
no longer true, so better use document title that is not likely
to change.

* TestWebKitAPI/Tests/WebKit2Gtk/TestInspectorServer.cpp:
(openRemoteDebuggingSession):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183543 268f45cc-cd09-0410-ab3c-d52691b4dbfc