Clean up: Remove unnecessary runtime computation of string length https://bugs.webkit.org/show_bug.cgi?id=144483
Reviewed by Joseph Pecoraro.

Following <http://trac.webkit.org/changeset/183649>, WebCore::fullyQualifiedInfoTableName()
computes strlen() of the first string literal as part of concatenating two string literals.
It is sufficient to use sizeof() - 1 instead of strlen() to compute the length of the first
string literal because the size of the string literal is known at compile time.

* Modules/webdatabase/DatabaseBackendBase.cpp:
(WebCore::fullyQualifiedInfoTableName):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183663 268f45cc-cd09-0410-ab3c-d52691b4dbfc