check-webkit-style fails due to system pylint https://bugs.webkit.org/show_bug.cgi?id=144360
Patch by Mario Sanchez Prada <mario@endlessm.com> on 2015-05-01
Reviewed by Darin Adler.

Make sure that the directory containing third party autoinstalled modules
is at the beginning of the search path, to avoid conflicts with other
versions of the same modules installed in the system (e.g. pylint).

* Scripts/webkitpy/thirdparty/__init__.py: Prepend the autoinstalled
modules directory to sys.path, instead of appending it.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183701 268f45cc-cd09-0410-ab3c-d52691b4dbfc