[ATK] AX: figure out platform difference for ATK to make accessibility/table-cell-display-block.html work https://bugs.webkit.org/show_bug.cgi?id=143083
Reviewed by Chris Fleizach.

Source/WebCore:

The main platform difference was that ATK was mapping the td element to
the accessible table cell, and the render block resulting from the style
wound up being the cell parent. In contrast, AX API was mapping the cell
element to an AXGroup which was a child of the AXCell. Rather than try to
bend yet another case to the will of our individual platforms' table needs,
it made more sense to clean up the code that is competing and move toward
more platform-unified handling. This is the first step in that direction,
with ATK moving closer toward the WebCore/AX API mapping.

No new tests. Instead, modified and fixed the test which was being skipped
for GTK, and also updated the expectations of some Mac tests whose FIXMEs
should now be fixed.

* accessibility/AccessibilityNodeObject.cpp:
(WebCore::AccessibilityNodeObject::supportsRequiredAttribute):
* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::determineAccessibilityRole):
* accessibility/AccessibilityTable.cpp:
(WebCore::AccessibilityTable::computeIsTableExposableThroughAccessibility):
* accessibility/AccessibilityTableCell.cpp:
(WebCore::AccessibilityTableCell::determineAccessibilityRole):
* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(createAccessibilityRoleMap):

LayoutTests:

* accessibility/roles-computedRoleString-expected.txt: Updated to reflect fixed FIXME.
* accessibility/roles-computedRoleString.html: Updated to reflect fixed FIXME.
* accessibility/table-cell-display-block-expected.txt: Modified to be more platform agnostic.
* accessibility/table-cell-display-block.html: Modified to be more platform agnostic.
* platform/efl/accessibility/table-detection-expected.txt: Updated to reflect mapping change.
* platform/gtk/TestExpectations: Unskip failing test from opening report.
* platform/gtk/accessibility/table-detection-expected.txt: Updated to reflect mapping change.
* platform/mac/accessibility/rowheader-outside-table-role-expected.txt: Updated to reflect mapping change.
* platform/mac/accessibility/rowheader-outside-table-role.html: Updated to reflect mapping change.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183798 268f45cc-cd09-0410-ab3c-d52691b4dbfc