ASSERTION when pasting text into the WebInspector console https://bugs.webkit.org/show_bug.cgi?id=144774
Reviewed by Ryosuke Niwa.

Fix assertion in didChangeSelection() meant to check that calling
editorState() does not cause a synchronous layout. The assertion
was not correct as it was relying on FrameView::needsLayout() and
we would hit it if calling editorState() would schedule a layout.
Instead, the new assertion relies on FrameView::layoutCount(),
which is more accurate.

* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::didChangeSelection):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@183969 268f45cc-cd09-0410-ab3c-d52691b4dbfc