[GTK] RunLoop constructor should properly retrieve or establish the thread-default GMainContext https://bugs.webkit.org/show_bug.cgi?id=144732
Reviewed by Carlos Garcia Campos.

RunLoop constructor in the GTK implementation should use the
existing thread-default context, create a new one if not on
the main thread, or use the global-default one if on the main
thread.

In RunLoop::run(), the GMainContext should then be pushed as
the thread-default before calling g_main_loop_run(), and popped
off when the main loop stops.

* wtf/gtk/RunLoopGtk.cpp:
(WTF::RunLoop::RunLoop):
(WTF::RunLoop::run):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184333 268f45cc-cd09-0410-ab3c-d52691b4dbfc