Simple line layout: Ignore -webkit-flow-*content while collecting text content for innerText. https://bugs.webkit.org/show_bug.cgi?id=145344 rdar://problem/20959522
Reviewed by Antti Koivisto.

This patch ensures that when a -webkit-flow-into content is present in a simple line layout container,
we skip it while collecting the text content for innerText.
It's not always guaranteed that a simple line layout container only contains simple text nodes. In such cases,
we need to ensure that text offset computing only takes the simple text nodes into account.

Source/WebCore:

Test: fast/text/simple-line-layout-inner-text-with-flow-content.html

* editing/TextIterator.cpp:
(WebCore::TextIterator::handleTextNode):
* editing/TextIterator.h:

LayoutTests:

* fast/text/simple-line-layout-inner-text-with-flow-content-expected.txt: Added.
* fast/text/simple-line-layout-inner-text-with-flow-content.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184825 268f45cc-cd09-0410-ab3c-d52691b4dbfc