AX: display:none content exposed to accessibility when aria-hidden is toggled on ancestor element https://bugs.webkit.org/show_bug.cgi?id=139142
Reviewed by Darin Adler.
Source/WebCore:

     
Amend the code that determines when an invisible, but aria-hidden=false, element is exposed to accessibility.

The new guideline is that you must have aria-hidden=false on every node that is not rendered (except text which inherits)
otherwise the element will not be visible.

Modified existing test: accessibility/aria-hidden-false-works-in-subtrees.html

* accessibility/AXObjectCache.cpp:
(WebCore::isNodeAriaVisible):
(WebCore::AXObjectCache::rootWebArea):

LayoutTests:


* accessibility/aria-hidden-false-works-in-subtrees.html:



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@184890 268f45cc-cd09-0410-ab3c-d52691b4dbfc