Web Inspector: Debugger Popover for Node object should have go-to-arrow to show the node in DOM tree https://bugs.webkit.org/show_bug.cgi?id=145594
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-06-03
Reviewed by Timothy Hatcher.

* UserInterface/Controllers/DOMTreeManager.js:
(WebInspector.DOMTreeManager.prototype.inspectElement):
Previously we could get an exception later on if we tried to
inspect an element that was not in the DOM. Bail if the node
is not in the DOM or not.

* UserInterface/Views/SourceCodeTextEditor.css:
(.popover .debugger-popover-content > .title > .go-to-arrow):
* UserInterface/Views/SourceCodeTextEditor.js:
(WebInspector.SourceCodeTextEditor.prototype._showPopoverForObject.data.pushNodeToFrontend):
(WebInspector.SourceCodeTextEditor.prototype._showPopoverForObject):
For a node type, add a go-to-arrow in the popover if it is in the DOM or not.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185150 268f45cc-cd09-0410-ab3c-d52691b4dbfc