REGRESSION (r184968): missing media player buttons (control bar exists, but no buttons) https://bugs.webkit.org/show_bug.cgi?id=145630
Reviewed by Zalan Bujtas.

Source/WebCore:

r184968 changed the initial value of m_intersectsCoverageRect to false. However,
this triggered a bug with mask layers, whose value of m_intersectsCoverageRect was never
updated.

Fix by copying the value of m_intersectsCoverageRect from the main layer to its
mask layer, just as we do for m_visibleRect and m_coverageRect.

Test: compositing/visible-rect/mask-layer-coverage.html

* platform/graphics/ca/GraphicsLayerCA.cpp:
(WebCore::GraphicsLayerCA::setVisibleAndCoverageRects):

LayoutTests:

Test that dumps coverage info for a layer with a mask.

* compositing/visible-rect/mask-layer-coverage-expected.txt: Added.
* compositing/visible-rect/mask-layer-coverage.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185187 268f45cc-cd09-0410-ab3c-d52691b4dbfc