DFG Object Allocation Sinking should not consider GetClosureVar as escapes https://bugs.webkit.org/show_bug.cgi?id=145904
Reviewed by Filip Pizlo.

The object allocation sinking phase is currently able to sink
CreateActivation nodes, but will consider any GetClosureVar node as
escaping.

This is not problematic in general as most of the GetClosureVar nodes
we would have been able to sink over will have been eliminated by CSE
anyway. Still, this is an oversight that we should fix since the
machinery is already in place.

* dfg/DFGObjectAllocationSinkingPhase.cpp:
(JSC::DFG::ObjectAllocationSinkingPhase::handleNode):
* dfg/DFGPromoteHeapAccess.h:
(JSC::DFG::promoteHeapAccess):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185505 268f45cc-cd09-0410-ab3c-d52691b4dbfc