WebKit crash while loading nytimes at JavaScriptCore: JSC::ExecutableAllocator::allocate + 276 https://bugs.webkit.org/show_bug.cgi?id=146163 <rdar://problem/20392986>
Reviewed by Michael Saboff.

There's no good way to test this in our test harness because we don't
have a way to simulate executable memory pressure, and doing so would
cause the cases that still use JITCompilationMustSucceed to crash.

Instead, I tested by manually forcing all regexp JIT compilation to
fail and running the JavaScriptCore tests.

* yarr/YarrJIT.cpp:
(JSC::Yarr::YarrGenerator::compile): Allow compilation to fail. We can
fall back to the regexp interpreter if we need to.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185770 268f45cc-cd09-0410-ab3c-d52691b4dbfc