Should reduce tile coverage for the first paint after a tab switch. <https://webkit.org/b/146252> <rdar://problem/19821583>
Reviewed by Darin Adler.

Reduce the number of tiles we need to paint after switching tabs,
to shorten the time it takes before we can flush pixels to screen.

We accomplish this by piggybacking on the "speculative tiling enabled"
mode of FrameView, which was previously only used to throttle painting
and layer flushes during page load.

When a FrameView becomes visible, which is what happens when you
switch to its tab, we revert the speculative tiling optimization to
its initial state, and reset the "scrolled by user" flag.

In practice this means that after switching tabs, we only generate
enough tiles to fill the viewport. Then, after 500ms has passed
or the user scrolls the page, we go back to the usual speculative
tiling mode.

* page/FrameView.cpp:
(WebCore::FrameView::show):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185897 268f45cc-cd09-0410-ab3c-d52691b4dbfc