Better fix for Element with blur backdrop-filter shows edge duplication and dark edges https://bugs.webkit.org/show_bug.cgi?id=146287 <rdar://problem/21530437>
Reviewed by Simon Fraser.

There is no need to carry around a bit in FilterOperations indicating what
type of source property the object came from. Whether or not to clip to the
bounds of the filter region is purely a property of the type of CALayer
we are applying the filter to.

This basically reverts r185846.

Covered by the test for the previous commit.

* css/CSSPropertyNames.in:
* css/StyleBuilderConverter.h:
(WebCore::StyleBuilderConverter::convertBackdropFilterOperations): Deleted.
* page/animation/CSSPropertyAnimation.cpp:
(WebCore::blendFilterOperations): Deleted.
* platform/graphics/ca/cocoa/PlatformCAFiltersCocoa.mm: Set the normalized
edges property if we are a CABackdropLayer.
* platform/graphics/filters/FilterOperations.cpp:
(WebCore::FilterOperations::operator==):
(WebCore::FilterOperations::operationsMatch):
(WebCore::FilterOperations::operator=): Deleted.
* platform/graphics/filters/FilterOperations.h:
(WebCore::FilterOperations::isUsedForBackdropFilters): Deleted.
(WebCore::FilterOperations::setUsedForBackdropFilters): Deleted.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@185925 268f45cc-cd09-0410-ab3c-d52691b4dbfc