Web Inspector: iOS 8: Uncaught Exception expanding Object Prototype https://bugs.webkit.org/show_bug.cgi?id=146427
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-06-30
Reviewed by Timothy Hatcher.

* UserInterface/Views/ObjectTreePropertyTreeElement.js:
(WebInspector.ObjectTreePropertyTreeElement.prototype._updateProperties):
* UserInterface/Views/ObjectTreeView.js:
(WebInspector.ObjectTreeView.prototype._updateProperties):
We are supposed to receive value properties for __proto__ properties
so that we can immediately start showing information about the property.
When that is not the case, which appeared to only be on legacy
backends for the top level prototype, just ignore it.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186120 268f45cc-cd09-0410-ab3c-d52691b4dbfc