Hide the inline controls when going into PiP https://bugs.webkit.org/show_bug.cgi?id=146487 <rdar://problem/19881159>
Reviewed by Eric Carlson.

When the presentation mode is PiP, toggle a class
on the container element so that we can completely
hide the control toolbar. The placard should still
remain visible.

* Modules/mediacontrols/mediaControlsiOS.css:
(video::-webkit-media-controls-panel-container.picture-in-picture): Add
a rule that hides the toolbar when PiP is active.
* Modules/mediacontrols/mediaControlsiOS.js:
(ControllerIOS.prototype.showControls):
(ControllerIOS.prototype.handlePresentationModeChange): Toggle a
class. Also fix a bug I noticed where exiting from PiP was
not auto-hiding the controls until the user tapped.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186183 268f45cc-cd09-0410-ab3c-d52691b4dbfc