[Win] REGRESSION (r185124) CACFLayer handling broken https://bugs.webkit.org/show_bug.cgi?id=146530 <rdar://problem/21642241>
Reviewed by Tim Horton.

At some point in the past, the set of LayerChange flags
overflowed the default MSVC enum type of 'int'. This caused
Windows rendering code to not receive various update notifications.

Zalan's change moved the DebugIndicatorsChanged flag into that
overflow set, which made it obvious that something was wrong.

The fix is to tell the compiler to use a compatible base type
for the enum.

* platform/graphics/ca/GraphicsLayerCA.h: Prevent overflow.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186210 268f45cc-cd09-0410-ab3c-d52691b4dbfc