REGRESSION (r183133-r183138): Secondary clicking in whitespace selects preceding word https://bugs.webkit.org/show_bug.cgi?id=146695 <rdar://problem/21441466>
Reviewed by Tim Horton.

When we would currently use dictionary lookup to decide on our text selection, first check if we
are targeting an editable field. If we are, use the original selection behavior. Otherwise, use
the new "dictionary lookup" selection.

* page/EventHandler.cpp:
(WebCore::EventHandler::selectClosestContextualWordOrLinkFromMouseEvent): Only use dictionary lookup
when our target node is not editable.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@186480 268f45cc-cd09-0410-ab3c-d52691b4dbfc